apply required changes
