+++ packages/expo-battery/android/src/main/java/expo/modules/battery/BatteryLevelReceiver.kt
@@ -6,6 +6,7 @@ import android.content.Intent
 import android.content.IntentFilter
 import android.os.BatteryManager
 import android.os.Bundle
+import android.util.Log
 import org.unimodules.core.interfaces.services.EventEmitter
 
 class BatteryLevelReceiver(private val eventEmitter: EventEmitter?) : BroadcastReceiver() {
@@ -19,7 +20,11 @@ class BatteryLevelReceiver(private val eventEmitter: EventEmitter?) : BroadcastR
 
   override fun onReceive(context: Context, intent: Intent) {
     val batteryIntent = context.applicationContext.registerReceiver(null, IntentFilter(Intent.ACTION_BATTERY_CHANGED))
-    val level = batteryIntent!!.getIntExtra(BatteryManager.EXTRA_LEVEL, -1)
+    if (batteryIntent == null) {
+      Log.e("Battery", "ACTION_BATTERY_CHANGED unavailable. Events wont be received")
+      return
+    }
+    val level = batteryIntent.getIntExtra(BatteryManager.EXTRA_LEVEL, -1)
     val scale = batteryIntent.getIntExtra(BatteryManager.EXTRA_SCALE, -1)
     val batteryLevel: Float = if (level != -1 && scale != -1) {
       level / scale.toFloat()
