+++ packages/expo-battery/android/src/main/java/expo/modules/battery/PowerSaverReceiver.kt
@@ -7,8 +7,9 @@ import android.os.Bundle
 import android.os.PowerManager
 import org.unimodules.core.interfaces.services.EventEmitter
 
+private val POWERMODE_EVENT_NAME = "Expo.powerModeDidChange"
+
 class PowerSaverReceiver(private val eventEmitter: EventEmitter?) : BroadcastReceiver() {
-  private val POWERMODE_EVENT_NAME = "Expo.powerModeDidChange"
 
   private fun onLowPowerModeChange(lowPowerMode: Boolean) {
     eventEmitter?.emit(POWERMODE_EVENT_NAME, Bundle().apply {
